添加灵药瓶药水效果

+++ src/main/java/com/pasterdream/pasterdreammod/world/block/dreamaccumulator/DreamAccumulatorRecipeSerializer.java
@@ -52,7 +52,7 @@ public class DreamAccumulatorRecipeSerializer implements RecipeSerializer<DreamA
 
         int time = buffer.readVarInt();
         ItemIngredient catalyst = ItemIngredient.fromNetwork(buffer);
-        float catalystMultiplier = buffer.readFloat();
+        double catalystMultiplier = buffer.readDouble();
         ItemStack output = buffer.readItem();
         return new DreamAccumulatorRecipe(id, biomeTags, time, catalyst, catalystMultiplier, output);
     }
